backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.1
firebase-admin==6.2.0
sentence-transformers==3.4.1
faiss-cpu==1.7.4
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0

